test: Move auth tests in separate folder
