+++ packages/testing/playwright/tests/ui/auth/authenticated.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../fixtures/base';
 
 test.describe('Authentication', () => {
 	const testCases = [
